Renamed theme folder
